requirements.txt
django
django-rest-framework
psycopg2
python-dateutil
django-countries
django-crispy-forms